added handler for shift+page up/down for ImageWindow partially fixed ProgressWindow bug

+++ trunk/ImageWindow.h
@@ -21,7 +21,7 @@ class ImageWindow:public Window
         void scrollImageY(int scrollMsg);
         void scrollImage(bool vert, int amount);
         
-        void zoomImage(int nlevels);
+        void zoomImage(float nlevels);
     	void onKeyDown(int virtualKey);    
 };
